Move [callbacks.js, loops.js] to utilities folder
